singleton instead of factory global, split out all functions for refactorablility
