+++ index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+// Copyright 2011 Chris Williams <chris@iterativedesigns.com>
+
+module.exports = require('./lib/factorysingleton').getInstance();
